chore(setup): Move forms-web-app source into /src and re-configure paths
