+++ services/forms-web-app/Dockerfile
@@ -7,7 +7,6 @@ ENV APP_NAME="${APP_NAME}"
 ENV BUILD_ID="${BUILD_ID}"
 ENV VERSION="${VERSION}"
 ENV PORT=3000
-ENV DEBUG=*
 # Do not rely on NODE_ENV - exists for performance reasons only
 ENV NODE_ENV=production
 ADD . .
