+++ services/forms-web-app/src/public/images/favicon.ico
